Refactor long press modal into separate component [RCORE-600]
